requirements.txt
langchain
langgraph
langchain-google-genai
wikipedia
python-dotenv
duckduckgo-search
requests